cleanup extensions
