+++ src/adapter/Extensions/ReaderWriterExtensions.cs
@@ -1,0 +1,60 @@
+using System;
+using System.IO;
+
+namespace NeoDebug
+{
+    static class ReaderWriterExtensions
+    {
+        public static ulong ReadVarInt(this BinaryReader reader, ulong max = ulong.MaxValue)
+        {
+            byte fb = reader.ReadByte();
+            ulong value;
+            if (fb == 0xFD)
+                value = reader.ReadUInt16();
+            else if (fb == 0xFE)
+                value = reader.ReadUInt32();
+            else if (fb == 0xFF)
+                value = reader.ReadUInt64();
+            else
+                value = fb;
+            if (value > max) throw new FormatException();
+            return value;
+        }
+
+        public static void WriteVarInt(this BinaryWriter writer, long value)
+        {
+            if (value < 0)
+                throw new ArgumentOutOfRangeException();
+            if (value < 0xFD)
+            {
+                writer.Write((byte)value);
+            }
+            else if (value <= 0xFFFF)
+            {
+                writer.Write((byte)0xFD);
+                writer.Write((ushort)value);
+            }
+            else if (value <= 0xFFFFFFFF)
+            {
+                writer.Write((byte)0xFE);
+                writer.Write((uint)value);
+            }
+            else
+            {
+                writer.Write((byte)0xFF);
+                writer.Write(value);
+            }
+        }
+
+        public static byte[] ReadVarBytes(this BinaryReader reader, int max = 0x1000000)
+        {
+            return reader.ReadBytes((int)reader.ReadVarInt((ulong)max));
+        }
+
+        public static void WriteVarBytes(this BinaryWriter writer, byte[] value)
+        {
+            writer.WriteVarInt(value.Length);
+            writer.Write(value);
+        }
+    }
+}
